+++ source/extensions/lightspeed.trex.stage_manager.plugin.filter.usd/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.3.0]
+### Added
+- Added `InstanceGroupFilterPlugin` filter plugin
+- Added `MeshGroupFilterPlugin` filter plugin
+### Changed
+- Renamed MeshPrimsFilterPlugin's label to "Geometry Prims"
+
 ## [2.2.1]
 ### Changed
 - Updated is_logic_graph filter plugin to be more inline with current style guide and practices