+++ source/extensions/omni.flux.stage_manager.plugin.filter.usd/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.6.0]
+### Changed
+- Updated `AdditionalFilterPlugin` to have headers for the different filter types
+- Added `InstanceGroupFilterPlugin` and `MeshGroupFilterPlugin`
+
 ## [2.5.0]
 ### Added
 - Added `VisiblePrimsFilterPlugin`