+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.11.0]
+### Added
+- Added `InstanceGroupInteractionPlugin` to the interactions plugins
+- Added `MeshGroupInteractionPlugin` to the interactions plugins
+
 ## [2.10.0]
 ### Added
 - Added `ParticleSystemsFilterPlugin` to the interactions plugins